requirements.txt
bottle==0.12.8
praw==2.1.21
requests==2.7.0
six==1.9.0
update-checker==0.11
wsgiref==0.1.2
ujson
gevent
gevent-socketio